requirements.txt
pydvdid>=1.0
requests>=2.9.1
xmltodict>=0.10.2
pyudev>=0.21.0
pyyaml>=3.12
flask>=1.0.2
flask-WTF>=0.14.2
flask-sqlalchemy>=2.3.2
flask-migrate>=2.2.1
flask-cors>=3.0.10
psutil>=5.4.6
robobrowser>=0.5.3
tinydownload>=0.1.0
netifaces>=0.10.9
flask-login>=0.5.0
apprise>=0.8.9
musicbrainzngs>=0.7.1
discid>=1.1.1
prettytable>=2.0.0
psutil
